fixup! JAMES-2082 Re-organize cassandra project hierarchy
